+++ library/src/actions/multipleOf/multipleOf.test.ts
@@ -1,11 +1,7 @@
 import { describe, expect, test } from 'vitest';
 import type { NumberIssue } from '../../schemas/index.ts';
 import { expectActionIssue, expectNoActionIssue } from '../../vitest/index.ts';
-import {
-  multipleOf,
-  type MultipleOfAction,
-  type MultipleOfIssue,
-} from './multipleOf.ts';
+import { multipleOf, type MultipleOfAction } from './multipleOf.ts';
 
 describe('multipleOf', () => {
   describe('should return action object', () => {
@@ -45,8 +41,6 @@ describe('multipleOf', () => {
   });
 
   describe('should return dataset without issues', () => {
-    const action = multipleOf(5);
-
     test('for untyped inputs', () => {
       const issues: [NumberIssue] = [
         {
@@ -59,7 +53,7 @@ describe('multipleOf', () => {
         },
       ];
       expect(
-        action['~run']({ typed: false, value: null, issues }, {})
+        multipleOf(5)['~run']({ typed: false, value: null, issues }, {})
       ).toStrictEqual({
         typed: false,
         value: null,
@@ -68,30 +62,52 @@ describe('multipleOf', () => {
     });
 
     test('for valid numbers', () => {
-      expectNoActionIssue(action, [-15, -10, -5, 0, 5, 10, 15]);
+      expectNoActionIssue(multipleOf(5), [-15, -10, -5, 0, 5, 10, 15]);
+    });
+
+    test('for valid bigints', () => {
+      expectNoActionIssue(multipleOf(5n), [-15n, -10n, -5n, 0n, 5n, 10n, 15n]);
     });
   });
 
   describe('should return dataset with issues', () => {
-    const action = multipleOf(5, 'message');
-    const baseIssue: Omit<MultipleOfIssue<number, 5>, 'input' | 'received'> = {
+    const baseIssue = {
       kind: 'validation',
       type: 'multiple_of',
       expected: '%5',
       message: 'message',
-      requirement: 5,
-    };
+    } as const;
 
     test('for invalid numbers', () => {
-      expectActionIssue(action, baseIssue, [-14, -9, -4, 1, 3, 6, 11]);
+      expectActionIssue(
+        multipleOf(5, 'message'),
+        { ...baseIssue, requirement: 5 },
+        [-14, -9, -4, 1, 3, 6, 11]
+      );
     });
 
     test('for infinity', () => {
-      expectActionIssue(action, baseIssue, [-Infinity, Infinity]);
+      expectActionIssue(
+        multipleOf(5, 'message'),
+        { ...baseIssue, requirement: 5 },
+        [-Infinity, Infinity]
+      );
     });
 
     test('for NaN', () => {
-      expectActionIssue(action, baseIssue, [NaN]);
+      expectActionIssue(
+        multipleOf(5, 'message'),
+        { ...baseIssue, requirement: 5 },
+        [NaN]
+      );
+    });
+
+    test('for invalid bigints', () => {
+      expectActionIssue(
+        multipleOf(5n, 'message'),
+        { ...baseIssue, requirement: 5n },
+        [-14n, -9n, -4n, 1n, 3n, 6n, 11n]
+      );
     });
   });
 });