+++ library/src/actions/regex/regex.ts
@@ -63,6 +63,8 @@ export interface RegexAction<
 /**
  * Creates a [regex](https://en.wikipedia.org/wiki/Regular_expression) validation action.
  *
+ * Hint: Be careful with the global flag `g` in your regex pattern, as it can lead to unexpected results. See [MDN](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp/test#using_test_on_a_regex_with_the_global_flag) for more information.
+ *
  * @param requirement The regex pattern.
  *
  * @returns A regex action.
@@ -73,6 +75,8 @@ export function regex<TInput extends string>(
 
 /**
  * Creates a [regex](https://en.wikipedia.org/wiki/Regular_expression) validation action.
+ *
+ * Hint: Be careful with the global flag `g` in your regex pattern, as it can lead to unexpected results. See [MDN](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp/test#using_test_on_a_regex_with_the_global_flag) for more information.
  *
  * @param requirement The regex pattern.
  * @param message The error message.