+++ library/src/regex.ts
@@ -17,7 +17,8 @@ export const CUID2_REGEX: RegExp = /^[a-z][\da-z]*$/u;
 /**
  * [Decimal](https://en.wikipedia.org/wiki/Decimal) regex.
  */
-export const DECIMAL_REGEX: RegExp = /^[+-]?\d+(?:\.\d+)?$/u;
+// eslint-disable-next-line redos-detector/no-unsafe-regex -- false positive
+export const DECIMAL_REGEX: RegExp = /^[+-]?(?:\d*\.)?\d+$/u;
 
 /**
  * [Digits](https://en.wikipedia.org/wiki/Numerical_digit) regex.