+++ library/src/schemas/variant/index.ts
@@ -1,3 +1,7 @@
-export * from './types.ts';
+export type {
+  VariantIssue,
+  VariantOptions,
+  VariantOptionsAsync,
+} from './types.ts';
 export * from './variant.ts';
 export * from './variantAsync.ts';