+++ packages/to-json-schema/src/utils/index.ts
@@ -1,1 +1,2 @@
+export * from './addError.ts';
 export * from './handleError.ts';